Refactor station name functionality to device name
- Renamed StationNameSerializer and StationNameView to DeviceNameSerializer and DeviceNameView, respectively, to better reflect their purpose.
- Updated API endpoint from /stations/name/ to /devices/name/ for consistency with device-related functionality.
- Added tests for the new device name endpoint to ensure correct functionality and error handling.
- Removed obsolete station tests to streamline the test suite.
